Move devcontainer config methods to a separate file
